updated global state for chat logic
