+++ client/src/components/CurrentChat/ChatMessage.jsx
@@ -7,8 +7,9 @@ const ChatMessage = ({ message }) => {
   const [editModal, setEditModal] = useState(false);
   const username = useRecoilValue(usernameState);
 
-  if(message.deleted) {
-     message.message_text = 'This message was deleted'
+  console.log('From inside ChatMessage: ', message)
+  if (message.deleted) {
+     //message.message_text = 'This message was deleted'
   }
   const editMessage = (e) => {
     e.preventDefault();
@@ -17,14 +18,15 @@ const ChatMessage = ({ message }) => {
   }
   const submitEditMessage = (e) => {
     e.preventDefault();
-    message.message_text = e.target.editText.value;
+    //this will make an axios call to update the text when message logic to the DB is working
+    //message.message_text = e.target.editText.value;
     setEditModal(false);
   }
   const deleteMessage = (e) => {
     e.preventDefault();
     if (confirm('Are you sure you want to delete this message?')) {
       console.log('Attempted to delete message')
-      message.message_text = 'This message was deleted'
+      //message.message_text = 'This message was deleted'
     }
   }
   const addMessageToTask = (e) => {
@@ -45,7 +47,7 @@ const ChatMessage = ({ message }) => {
           <button className='current-chat-edit-button' onClick={editMessage}>Edit</button>
           <button className='current-chat-delete-button' onClick={deleteMessage}>Delete</button>
           <button className='current-chat-add-task-button' onClick={addMessageToTask}>Add to task</button>
-        <div>Date sent/created</div>
+          <div>Date sent/created: {message.created}</div>
         </div>
       </div>
     )
