+++ client/src/components/userAtoms.js
@@ -7,8 +7,13 @@ export const usernameState = atom({
 });
 
 export const groupState = atom({
-  key: 'gropState',
+  key: 'groupState',
+  default: 'general',
+});
+
+export const messageState = atom({
+  key: 'messageState',
   // Change default when done testing
-  default: null,
-})
+  default: [],
+});
